Move find/hide/explosion handlers to GameMode
